requirements.txt
black==22.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
cssselect==1.1.0
greenlet==1.1.2
idna==3.3
lxml==4.8.0
mypy-extensions==0.4.3
parsel==1.6.0
pathspec==0.9.0
platformdirs==2.5.1
psycopg2-binary==2.9.3
python-decouple==3.6
requests==2.27.1
six==1.16.0
tomli==2.0.1
typing_extensions==4.1.1
urllib3==1.26.9
w3lib==1.22.0
